// simulator-spa/src/app/core/simulator/simulator.types.ts
export interface CarData
{
  turnSignalLeft: boolean;
  turnSignalRight: boolean;
  positionLight: boolean;
  cruiseLight: boolean;
  fullheadLight: boolean;
  motor: boolean;
  battery:  boolean;
  handbrake: boolean;
  airConditioner: boolean;
  airSpeedFan: number;
  airTemperature: number;
  carTemperature: number;
}